- windoof development
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@472 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/FormatInputStream.cpp
@@ -55,7 +55,7 @@ namespace org {
                 return 0;
             }
 
-            long FormatInputStream::getFileSize() {
+            long long int FormatInputStream::getFileSize() {
                 return formatCtx->file_size;
             }
